+++ cmd/hidden_lake/loader/pkg/app/app_test.go
@@ -1,0 +1,5 @@
+package app
+
+import "testing"
+
+func TestNothing(t *testing.T) {}