+++ cmd/hidden_lake/loader/pkg/settings/settings.go
@@ -1,0 +1,15 @@
+package settings
+
+const (
+	CServiceName  = "HLL"
+	CTitlePattern = "go-peer/hidden-lake-loader"
+)
+
+const (
+	CPathCFG = "hll.cfg"
+)
+
+const (
+	CDefaultConsumerAddress = "127.0.0.1:9582"
+	CDefaultProducerAddress = "127.0.0.2:9582"
+)