+++ examples/echo_service/prod_test/gen_cfgs/main.go
@@ -90,9 +90,9 @@ func initRecvNode(pConnects string) *sNodeHLS {
 		"work_size_bits": 20,
 		"key_size_bits": 4096,
 		"queue_period_ms": 5000,
-		"limit_void_size_bytes": 4096
+		"limit_void_size_bytes": 4096,
+		"network_key": "%s"
 	},
-	"network_key": "%s",
 	"logging": ["info", "warn", "erro"],
 	"services": {
 		"hidden-echo-service": "localhost:8080"
@@ -120,9 +120,9 @@ func initSendNode(pConnects string) *sNodeHLS {
 		"work_size_bits": 20,
 		"key_size_bits": 4096,
 		"queue_period_ms": 5000,
-		"limit_void_size_bytes": 4096
+		"limit_void_size_bytes": 4096,
+		"network_key": "%s"
 	},
-	"network_key": "%s",
 	"logging": ["info", "warn", "erro"],
 	"address": {
 		"http": "localhost:7572"
@@ -161,9 +161,9 @@ clean:
 		"work_size_bits": 20,
 		"key_size_bits": 4096,
 		"queue_period_ms": 5000,
-		"limit_void_size_bytes": 4096
+		"limit_void_size_bytes": 4096,
+		"network_key": "%s"
 	},
-	"network_key": "%s",
 	"connections": [%s]
 }
 `,